test: split migration test harness helpers
